Prototyping 06
New proxy mode. In proxy mode, the data can be transferred to an external web server and remain encrypted there. Proxy mode must also be activated in the Pi.Alert Master. So that it does not search for the data locally, but retrieves it from the proxy.

What currently works:

- Scans
- Arpscan and Fritzbox import have been tested
- Mikrotik and UniFy should work, but could not be tested
- A json is generated and encrypted using openssl
- The API is called and the data is transmitted
- Receiving and decrypting the data on the API side

What does not work:
- The entire processing of the data
- the configuration of a satellite in the frontend
- installation process

Current status:
- Still Prototype
- Continuation of the project cannot currently be guaranteed